Comments in aliens and ansi

+++ SpaceProgram/inc/aliens.h
@@ -1,12 +1,22 @@
 #ifndef ALIENS_H
 #define ALIENS_H
 
-//#include "stm32f30x_rcc.h"
 #include "ansi.h"
 #include "30010_io.h"
 #include "Landscape.h"
 #include "SpaceShip.h"
 
+
+/*
+ * Struct: alien_t
+ * --------------------------
+ * A struct to represent an alien
+ *
+ * posX: x-coordinate center position of the alien
+ * posY: y-coordinate center position of the alien
+ * dir: the direction the alien i moving. At 1 it moves right and at 2 it moves left
+ *
+ */
 typedef struct {
     int posX, posY;
     int8_t dir;
@@ -18,7 +28,7 @@ void updateAliens(alien_t aliens[], SpaceShip_t * ship, int8_t numAliens, uint8_
 void alienKilled(alien_t aliens[], int8_t index, int8_t numAliens);
 int8_t spawnAlien(alien_t aliens[], int8_t emptyIndex, int spawnRate);
 void makeAlien(alien_t aliens[], int8_t emptyIndex);
-int8_t collide(alien_t aliens[], SpaceShip_t *ship, int numAliens, uint8_t *buffer);
+int8_t collide(alien_t aliens[], SpaceShip_t *ship, int8_t numAliens, uint8_t *buffer);
 
 
 #endif
